Test binary search and fix search in histogram
When an histogram is specified, a mapping between value to pixel is generated
and stored into an array. The value is binary searched in the array and the
corresponding pixel is returned. There was a bug in this procedure: always the
rightmost value was returned for intermediate values searched.

For example, searching 5.1 and 5.9 in:

[1, 5, 6, 56]

would have returned element at index 2 (i.e., 6) in both cases.

Now the closest element is returned. In the example above, the element at index
1 would be returned when 5.1 is searched, and the element at index 2 would be
returned when 5.9 is searched.