Set width and offset of slider bar correctly
There was a miscalculation on the width of the slider bar, which now is set
correctly. The bar must span from he first pixel of the left handle to the
last pixel of the last handle. In case of single bar, the naked end of the bar
should be set precisely at the pixel value.